Cleanup / consolidate unit testing code
- Move testing code out of helpers.py
- Create new file unit_test.py
