+++ InvenTree/InvenTree/helpers.py
@@ -8,18 +8,15 @@ import os
 import os.path
 import re
 from decimal import Decimal, InvalidOperation
-from pathlib import Path
 from wsgiref.util import FileWrapper
 
 from django.conf import settings
-from django.contrib.auth.models import Permission
 from django.contrib.staticfiles.storage import StaticFilesStorage
 from django.core.exceptions import FieldError, ValidationError
 from django.core.files.storage import default_storage
 from django.core.validators import URLValidator
 from django.db.utils import OperationalError, ProgrammingError
 from django.http import StreamingHttpResponse
-from django.test import TestCase
 from django.utils.translation import gettext_lazy as _
 
 import moneyed.localization
@@ -36,7 +33,6 @@ from common.notifications import (InvenTreeNotificationBodies,
                                   NotificationBody, trigger_notification)
 from common.settings import currency_code_default
 
-from .api_tester import ExchangeRateMixin, UserMixin
 from .settings import MEDIA_URL, STATIC_URL
 
 logger = logging.getLogger('inventree')
@@ -850,75 +846,6 @@ def validateFilterString(value, model=None):
     return results
 
 
-def addUserPermission(user, permission):
-    """Shortcut function for adding a certain permission to a user."""
-    perm = Permission.objects.get(codename=permission)
-    user.user_permissions.add(perm)
-
-
-def addUserPermissions(user, permissions):
-    """Shortcut function for adding multiple permissions to a user."""
-    for permission in permissions:
-        addUserPermission(user, permission)
-
-
-def getMigrationFileNames(app):
-    """Return a list of all migration filenames for provided app."""
-    local_dir = Path(__file__).parent
-    files = local_dir.joinpath('..', app, 'migrations').iterdir()
-
-    # Regex pattern for migration files
-    regex = re.compile(r"^[\d]+_.*\.py$")
-
-    migration_files = []
-
-    for f in files:
-        if regex.match(f.name):
-            migration_files.append(f.name)
-
-    return migration_files
-
-
-def getOldestMigrationFile(app, exclude_extension=True, ignore_initial=True):
-    """Return the filename associated with the oldest migration."""
-    oldest_num = -1
-    oldest_file = None
-
-    for f in getMigrationFileNames(app):
-
-        if ignore_initial and f.startswith('0001_initial'):
-            continue
-
-        num = int(f.split('_')[0])
-
-        if oldest_file is None or num < oldest_num:
-            oldest_num = num
-            oldest_file = f
-
-    if exclude_extension:
-        oldest_file = oldest_file.replace('.py', '')
-
-    return oldest_file
-
-
-def getNewestMigrationFile(app, exclude_extension=True):
-    """Return the filename associated with the newest migration."""
-    newest_file = None
-    newest_num = -1
-
-    for f in getMigrationFileNames(app):
-        num = int(f.split('_')[0])
-
-        if newest_file is None or num > newest_num:
-            newest_num = num
-            newest_file = f
-
-    if exclude_extension:
-        newest_file = newest_file.replace('.py', '')
-
-    return newest_file
-
-
 def clean_decimal(number):
     """Clean-up decimal value."""
     # Check if empty
@@ -1089,11 +1016,6 @@ def inheritors(cls):
                 subcls.add(child)
                 work.append(child)
     return subcls
-
-
-class InvenTreeTestCase(ExchangeRateMixin, UserMixin, TestCase):
-    """Testcase with user setup buildin."""
-    pass
 
 
 def notify_responsible(instance, sender, content: NotificationBody = InvenTreeNotificationBodies.NewOrder, exclude=None):
