+++ InvenTree/InvenTree/unit_test.py
@@ -1,18 +1,89 @@
-"""Helper functions for performing API unit tests."""
+"""Helper functions for unit testing / CI"""
 
 import csv
 import io
 import re
+from pathlib import Path
 
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group
+from django.contrib.auth.models import Group, Permission
 from django.http.response import StreamingHttpResponse
+from django.test import TestCase
 
 from djmoney.contrib.exchange.models import ExchangeBackend, Rate
 from rest_framework.test import APITestCase
 
 from plugin import registry
 from plugin.models import PluginConfig
+
+
+def addUserPermission(user, permission):
+    """Shortcut function for adding a certain permission to a user."""
+    perm = Permission.objects.get(codename=permission)
+    user.user_permissions.add(perm)
+
+
+def addUserPermissions(user, permissions):
+    """Shortcut function for adding multiple permissions to a user."""
+    for permission in permissions:
+        addUserPermission(user, permission)
+
+
+def getMigrationFileNames(app):
+    """Return a list of all migration filenames for provided app."""
+    local_dir = Path(__file__).parent
+    files = local_dir.joinpath('..', app, 'migrations').iterdir()
+
+    # Regex pattern for migration files
+    regex = re.compile(r"^[\d]+_.*\.py$")
+
+    migration_files = []
+
+    for f in files:
+        if regex.match(f.name):
+            migration_files.append(f.name)
+
+    return migration_files
+
+
+def getOldestMigrationFile(app, exclude_extension=True, ignore_initial=True):
+    """Return the filename associated with the oldest migration."""
+    oldest_num = -1
+    oldest_file = None
+
+    for f in getMigrationFileNames(app):
+
+        if ignore_initial and f.startswith('0001_initial'):
+            continue
+
+        num = int(f.split('_')[0])
+
+        if oldest_file is None or num < oldest_num:
+            oldest_num = num
+            oldest_file = f
+
+    if exclude_extension:
+        oldest_file = oldest_file.replace('.py', '')
+
+    return oldest_file
+
+
+def getNewestMigrationFile(app, exclude_extension=True):
+    """Return the filename associated with the newest migration."""
+    newest_file = None
+    newest_num = -1
+
+    for f in getMigrationFileNames(app):
+        num = int(f.split('_')[0])
+
+        if newest_file is None or num > newest_num:
+            newest_num = num
+            newest_file = f
+
+    if exclude_extension:
+        newest_file = newest_file.replace('.py', '')
+
+    return newest_file
 
 
 class UserMixin:
@@ -337,3 +408,8 @@ class InvenTreeAPITestCase(ExchangeRateMixin, UserMixin, APITestCase):
             data.append(entry)
 
         return data
+
+
+class InvenTreeTestCase(ExchangeRateMixin, UserMixin, TestCase):
+    """Testcase with user setup buildin."""
+    pass
