+++ InvenTree/company/test_api.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 
 from rest_framework import status
 
-from InvenTree.api_tester import InvenTreeAPITestCase
+from InvenTree.unit_test import InvenTreeAPITestCase
 
 from .models import Company, Contact, SupplierPart
