+++ InvenTree/company/test_views.py
@@ -2,7 +2,7 @@
 
 from django.urls import reverse
 
-from InvenTree.helpers import InvenTreeTestCase
+from InvenTree.unit_test import InvenTreeTestCase
 
 
 class CompanyViewTest(InvenTreeTestCase):
