+++ InvenTree/part/test_bom_export.py
@@ -4,7 +4,7 @@ import csv
 
 from django.urls import reverse
 
-from InvenTree.helpers import InvenTreeTestCase
+from InvenTree.unit_test import InvenTreeTestCase
 
 
 class BomExportTest(InvenTreeTestCase):
