+++ InvenTree/plugin/base/barcodes/test_barcode.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 
 from rest_framework import status
 
-from InvenTree.api_tester import InvenTreeAPITestCase
+from InvenTree.unit_test import InvenTreeAPITestCase
 from stock.models import StockItem
