+++ InvenTree/stock/tests.py
@@ -9,8 +9,8 @@ from django.test import override_settings
 from build.models import Build
 from common.models import InvenTreeSetting
 from company.models import Company
-from InvenTree.helpers import InvenTreeTestCase
 from InvenTree.status_codes import StockHistoryCode
+from InvenTree.unit_test import InvenTreeTestCase
 from order.models import SalesOrder
 from part.models import Part
