+++ InvenTree/users/test_api.py
@@ -3,7 +3,7 @@
 from django.contrib.auth.models import Group, User
 from django.urls import reverse
 
-from InvenTree.api_tester import InvenTreeAPITestCase
+from InvenTree.unit_test import InvenTreeAPITestCase
 
 
 class UserAPITests(InvenTreeAPITestCase):
